move pass_index
