+++ SIRF_data_preparation/evaluation_utilities.py
@@ -5,7 +5,6 @@ from typing import Iterable
 
 import matplotlib.pyplot as plt
 import numpy as np
-from scipy.ndimage import binary_erosion
 
 import sirf.STIR as STIR
 from petric import QualityMetrics
@@ -23,21 +22,6 @@ def get_metrics(qm: QualityMetrics, iters: Iterable[int], srcdir='.'):
     """Read 'iter_{iter_glob}.hv' images from datadir, compute metrics and return as 2d array"""
     return np.asarray([
         list(qm.evaluate(STIR.ImageData(str(Path(srcdir) / f'iter_{i:04d}.hv'))).values()) for i in iters])
-
-
-def pass_index(metrics: np.ndarray, thresh: Iterable, window: int = 10) -> int:
-    """
-    Returns first index of `metrics` with value <= `thresh`.
-    The values must remain below the respective thresholds for at least `window` number of entries.
-    Otherwise raises IndexError.
-    """
-    thr_arr = np.asanyarray(thresh)
-    assert metrics.ndim == 2
-    assert thr_arr.ndim == 1
-    assert metrics.shape[1] == thr_arr.shape[0]
-    passed = (metrics <= thr_arr[None]).all(axis=1)
-    res = binary_erosion(passed, structure=np.ones(window), origin=-(window // 2))
-    return np.where(res)[0][0]
 
 
 def plot_metrics(iters: Iterable[int], m: np.ndarray, labels=None, suffix=""):
